# coordinator/flex/server/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: graphscope@alibaba-inc.com
    name: GraphScope
  description: |-
    This is a specification for GraphScope FLEX HTTP service based on the OpenAPI 3.0 specification. You can find out more details about
    specification at [doc](https://swagger.io/specification/v3/).

    Some useful links:
    - [GraphScope Repository](https://github.com/alibaba/GraphScope)
    - [The Source API definition for GraphScope Interactive](https://github.com/GraphScope/portal/tree/main/httpservice)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GraphScope FLEX HTTP SERVICE API
  version: 1.9.2
...
servers:
- url: /
...
paths:
...
  /api/v1/graph/{graph_id}/dataloading:
    post:
      description: Submit a dataloading job
      operationId: submit_dataloading_job
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataloadingJobConfig'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDataloadingJobResponse'
          description: Successfully submitted the job
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
  /api/v1/graph/{graph_id}/dataloading/config:
    post:
      description: Post to get the data loading configuration for MapReduce Task
      operationId: get_dataloading_job_config
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataloadingJobConfig'
        required: true
      responses:
        "200":
          content:
            application/json:
              example: |-
                split.size=32
                separator=\\|
                graph.endpoint=11.11.11.11:55556
                project=graphscope
              schema:
                $ref: '#/components/schemas/DataloadingMRJobConfig'
          description: Successfully returned the job configuration for MapReduce Task
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
...
  /api/v1/job:
    get:
      description: List all jobs
      operationId: list_jobs
      responses:
        "200":
          content:
            application/json:
              example:
              - id: JOB-QWERTYUIOPASDFGH-1
                type: DATA IMPORT
                status: RUNNING
                start_time: 2024-02-24 00:00:00
                log: logging...
                detail:
                  graph_id: GRAPHID
              - id: JOB-QWERTYUIOPASDFGH-2
                type: DATA IMPORT
                status: CANCELLED
                start_time: 2024-02-24 00:00:00
                log: logging...
                detail:
                  graph_id: GRAPHID
              - id: JOB-QWERTYUIOPASDFGH-3
                type: DATA IMPORT
                status: SUCCESS
                start_time: 2024-02-24 00:00:00
                end_time: 2024-02-24 12:00:00
                detail:
                  graph_id: GRAPHID
              - id: JOB-QWERTYUIOPASDFGH-4
                type: DATA IMPORT
                status: FAILED
                start_time: 2024-02-24 00:00:00
                end_time: 2024-02-24 12:00:00
                log: logging...
                detail:
                  graph_id: GRAPHID
              - id: JOB-QWERTYUIOPASDFGH-5
                type: DATA IMPORT
                status: WAITING
                detail:
                  graph_id: GRAPHID
              schema:
                items:
                  $ref: '#/components/schemas/JobStatus'
                type: array
          description: Successful returned all the jobs
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
  /api/v1/job/{job_id}:
    delete:
      description: Delete job by ID
      operationId: delete_job_by_id
      parameters:
      - explode: false
        in: path
        name: job_id
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: delete_scheduler
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successfuly cancelled the job
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
    get:
      description: Get job status by ID
      operationId: get_job_by_id
      parameters:
      - explode: false
        in: path
        name: job_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Successfully returned the job status
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
  /api/v1/job/{job_id}/pause:
    post:
      description: Pause an existing job
      operationId: pause_job
      parameters:
      - explode: false
        in: path
        name: job_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successfully paused the job
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
  /api/v1/job/{job_id}/resume:
    post:
      description: Resume an existing job
      operationId: resume_job
      parameters:
      - explode: false
        in: path
        name: job_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successfully resumed a job
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Job
      x-openapi-router-controller: flex.server.controllers.job_controller
...
components:
...
  schemas:
    APIResponse:
      type: string
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          description: Error code
          format: int32
          title: code
          type: integer
        message:
          description: Error message
          title: message
          type: string
      required:
      - code
      - message
      title: Error
...
    DataloadingJobConfig:
      example:
        schedule: schedule
        loading_config:
          format:
            metadata:
              key: ""
            type: type
          import_option: overwrite
        vertices:
        - type_name: type_name
        - type_name: type_name
        repeat: once
        edges:
        - type_name: type_name
          source_vertex: source_vertex
          destination_vertex: destination_vertex
        - type_name: type_name
          source_vertex: source_vertex
          destination_vertex: destination_vertex
      properties:
        loading_config:
          $ref: '#/components/schemas/DataloadingJobConfig_loading_config'
        vertices:
          items:
            $ref: '#/components/schemas/DataloadingJobConfig_vertices_inner'
          title: vertices
          type: array
        edges:
          items:
            $ref: '#/components/schemas/DataloadingJobConfig_edges_inner'
          title: edges
          type: array
        schedule:
          description: format with '2023-02-21 11:56:30'
          nullable: true
          title: schedule
          type: string
        repeat:
          enum:
          - once
          - day
          - week
          nullable: true
          title: repeat
          type: string
      required:
      - edges
      - loading_config
      - vertices
      title: DataloadingJobConfig
      x-body-name: dataloading_job_config
    DataloadingMRJobConfig:
      example:
        config: config
      properties:
        config:
          title: config
          type: string
      required:
      - config
      title: DataloadingMRJobConfig
      x-body-name: dataloading_mr_job_config
    JobStatus:
      example:
        start_time: start_time
        log: log
        end_time: end_time
        id: id
        detail:
          key: ""
        type: type
        status: RUNNING
      properties:
        id:
          title: id
          type: string
        type:
          title: type
          type: string
        status:
          enum:
          - RUNNING
          - SUCCESS
          - FAILED
          - CANCELLED
          - WAITING
          - PAUSED
          title: status
          type: string
        start_time:
          title: start_time
          type: string
        end_time:
          title: end_time
          type: string
        log:
          description: logview URL or log string
          title: log
          type: string
        detail:
          additionalProperties: true
          title: detail
          type: object
      required:
      - id
      - status
      - type
      title: JobStatus
    CreateDataloadingJobResponse:
      example:
        job_id: job_id
      properties:
        job_id:
          title: job_id
          type: string
      required:
      - job_id
      title: CreateDataloadingJobResponse
...
    DataloadingJobConfig_loading_config_format:
      example:
        metadata:
          key: ""
        type: type
      properties:
        type:
          title: type
          type: string
        metadata:
          additionalProperties: true
          title: metadata
          type: object
      title: DataloadingJobConfig_loading_config_format
    DataloadingJobConfig_loading_config:
      example:
        format:
          metadata:
            key: ""
          type: type
        import_option: overwrite
      properties:
        import_option:
          enum:
          - init
          - overwrite
          example: overwrite
          title: import_option
          type: string
        format:
          $ref: '#/components/schemas/DataloadingJobConfig_loading_config_format'
      title: DataloadingJobConfig_loading_config
    DataloadingJobConfig_vertices_inner:
      example:
        type_name: type_name
      properties:
        type_name:
          title: type_name
          type: string
      title: DataloadingJobConfig_vertices_inner
      type: object
    DataloadingJobConfig_edges_inner:
      example:
        type_name: type_name
        source_vertex: source_vertex
        destination_vertex: destination_vertex
      properties:
        type_name:
          title: type_name
          type: string
        source_vertex:
          title: source_vertex
          type: string
        destination_vertex:
          title: destination_vertex
          type: string
      title: DataloadingJobConfig_edges_inner
      type: object